fold AffineGapPenaltyAlignment into ReadAlignment
AffineGapPenaltyAlignment was a standalone object with one public
method (in spirit; scoreAlignmentPaths was public but shouldn't have
been. I made it package-private here so that tests can access it) that
was the main way of creating a ReadAlignment; arguably that makes it
just a ReadAlignment constructor, so I've folded it into ReadAlignment
accordingly.
