+++ src/main/scala/org/hammerlab/guacamole/commands/GermlineAssemblyCaller.scala
@@ -4,7 +4,7 @@ import breeze.linalg.DenseVector
 import breeze.stats.{mean, median}
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
-import org.hammerlab.guacamole.alignment.AffineGapPenaltyAlignment
+import org.hammerlab.guacamole.alignment.ReadAlignment
 import org.hammerlab.guacamole.assembly.{AssemblyArgs, AssemblyUtils}
 import org.hammerlab.guacamole.distributed.WindowFlatMapUtils.windowFlatMapWithState
 import org.hammerlab.guacamole.likelihood.Likelihood
@@ -200,7 +200,7 @@ object GermlineAssemblyCaller {
                       path,
                       referenceStart,
                       referenceContig,
-                      path => AffineGapPenaltyAlignment.align(path, currentReference),
+                      ReadAlignment(_, currentReference),
                       buildVariant
                     )
                   )
